requirements.txt
flask==3.0.3
gunicorn==22.0.0
chromadb==0.5.3
google-cloud-aiplatform==1.56.0
google-cloud-secret-manager==2.20.0
google-cloud-sql-connector==1.13.0
sqlalchemy==2.0.35
pymysql==1.1.1
langchain==0.3.1
langchain-community==0.3.1
langchain-google-vertexai==0.1.2
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3